plugins.virt_test: FIX: Use defaults instead of None
When adding cmdline arguments, setting is read to find default values,
but they are used only for help, not for the real defaults. The code
underneath uses other methods to obtain the value than settings in case
of None value making it impossible to use the settings.

This patch uses the settings values as defaults instead the None value.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>