avocado.core.plugins.virt_test: Return empty test list on LexerError
The way the virt plugin selects its tests is by passing
the user provided URL to the cartesian parser. Some of
the identifiers the user might pass are illegal in this
context (example, examples/tests/sleeptest.py contains
slashes, that are illegal for the cartesian parser).

The fact the identifier is illegal for the cartesian
parser already means it's invalid, no need to propagate
exceptions. Just return an empty list and let other
test loader plugins to try to identify it.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>