avocado.core.plugins.virt_test: Don't sys.exit in plugin
virt_test is a plugin and in case of error it should only raise a sane
exception and leave the decision on the code below instead of sys.exit.

As a side effect this allows the loader plugin to log the failure and
proceed with other loader plugins with the discovery while still
exiting when the url is not resolved by later loader plugins.

Another thing that wasn't clear after many changes of the code base
was from where each invalid value was coming, let's also point out
more clearly that information from the user, in order to help with
fixing the problems promptly.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>
Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>