plugins.virt_test: Don't sys.exit in plugin
virt_test is a plugin and in case of error it should only raise a sane
exception and leave the decision on the code below instead of sys.exit.

As a side effect this allows the loader plugin to log the failure and
proceed with other loader plugins with the discovery while still
exiting when the url is not resolved by later loader plugins.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>