avocado-vt: Another fix for the empty logs
It was pointed out that using the API available
to test writers is a layer violation (plugins are
supposed to be 'core' code), so let's have a different
fix -> one of the refactors of the test class turned
the .start_logging and .stop_logging methods 'private'.

The fix that keeps the boundaries intact involves
renaming the methods in the VirtTest class implementation
to have the same names as the parent class -> this will
make them to be executed, logging calls that go to the
root logger also get logged in the test log.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>