virttest/virt_vm.py: Adjust timing to get mac list
The fef5c4c move mac_list() to the beginning of get_address(),
which leads to verification of arping result failed until
the timeout to be zero.

The reason as following,
The self.virtnet of VM is created by VirtNet.__init__,
and the self.virtnet.mac isn't set any value.
So it's better to obtain the mac list after checking the
existence of this attribute by nic.has_key('mac').