virttest.video_maker: Add gstreamer 1.0 video encoding class
In recent distros, such as Fedora 20 and later, when you
try to load the python-gstreamer bindings, the program
using it will print:

** Message: pygobject_register_sinkfunc is deprecated (GstObject)

If we can help it, such messages shouldn't be printed. So
let's introduce a new video encoder, GiEncoder, that is
capable of using the gobject-introspection gstreamer
bindings. Do a little refactor and make the interfaces
of each encoder equivalent, and change the name of the
'start' method to 'encode' for more clarity. Adjust env_process
accordingly.

With this, older distros will be able to use the python-gstreamer,
and newer distros will use the recent, non deprecated interface.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>