virttest.env_process: Move virt_video import next to its usage
Loading the virt_video top level in avocado is causing
an unexpected side effect in avocado, if the system in
question uses the gstreamer-python bindings -> the
initialization of the library is causing the virt plugin
help messages to not be displayed. Let's move the import
to the place where it's actually used, solving this problem.

This bug doesn't affect newer distros, where we switched
to gi (gobject introspection).

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>