virttest.bootstrap: Force backend dir synchronization
Sometimes due to a setup problem (usually the user
interrupts the bootstrap or some other condition that
leads to an incomplete local config set) we might
get plugins not working.

Let's copy the contents of the base directory over the
local config set whenever --vt-bootstrap is called using
distutils, so we keep things up to date and consistent,
avoiding configuration based problems.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>