virttest.storage: Secure the backup handling
Currently when avocado is interrupted while creating backup, partial
backup is left on the drive, which can corrupt the testing. This patch
creates the backup into ".part" file and then renames it to the target
name. Rename is atomic so either the target file remains untouched and
.part file contains incomplete backup, or full backup is performed and
present in the target file.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>