vt: Remove --vt-setup
This patch removes the --vt-setup, because the usage is a bit
controversial. The only purpose is to uncompress the asset file and
it does it for every job (when specified).

Instead people should use --vt-bootstrap to obtain asset file and
default image backups (.backup files) which are performed automatically
(unless disabled).

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>