shared.cfg.base: Change the default for migrate_vms
The "base.cfg" is processed as first, thus "migrate_vms" always contains
only the "vms" from "base.cfg". This is not true for all tests that
modifies the "vms" in following cfg files (eg. "subtests.cfg"). This
leads in additional vms being fired up during preprocess for all cases
where vms doesn't contain the machines defined in "base.cfg" (usually
when vms = "").

Good think about "migrate_vms" is, that it extends the list of vms of
the actual "vms" value on-the-fly during env_preprocess, thus there is no
need to inject "vms" in the first place. And this is what this patch
actually does, it removes the default and leaves it up to the
env_preprocess to join vms to the migrate_vms value set by other cfgs.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>