avocado-vt: Fix unattended install
A long time ago, back in 2012 (see commit
feaa5e5a21733d4d1568e08879974ac55ef7bbad), unattended
install started to use the serial log file contents
instead of the serial console aexpect session. The
change was motivated by unattended install tests
with ping pong migration, that made the finish
message to be lost because of the switch between
src and dst.

Later on, we unified libvirt and qemu unattended
install mechanism, by making both read from the
log. Unfortunately, keeping up both implementations
working simultaneously was difficult, given that
people working with libvirt rarely tested on qemu
and vice-versa.

So here we go again. The QEMU implementation got
broken by patch 319887f26ada824033fde48812aa7715a0f4317e,
by yours truly, made when working on the libvirt
subtest bugs.

This time, let's introduce a new attribute to the
VM objects in both qemu and libvirt VM classes:
vm.serial_console_log. This reflects the actual log
file, and not the socket name of the serial console
device. We made it so that it works on both the
QEMU and Libvirt VM implementations. This time around,
both implementations were tested. For the future, we
have to be more careful with changes in this code,
and require from patch senders that both backends are
tested.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>