utils_misc: Deprecate find_command to avoid duplicate
The find_command in utils_misc is already implemented in
avocado.utils.path.

This command deprecate this function and show a warning logging
if some command still using it.
We can't change the existing reference of this function in
virt-test providers, since it will fail if someone still using
virt-test.

Also some reference to path.command is changed to path.find_command
since the command function no longer exists.

Signed-off-by: Hao Liu <hliu@redhat.com>