virttest.utils_env: Remove env file upon garbage collection
In the early days of KVM autotest, the environment file
was conceived as a way to keep virtual machines running
between autotest executions. Time has passed, and we
realized that this behavior is not the best, as it may
consume resources, and our users might be oblivious to
the fact the vms are running on the first place. It's
fine to have the environment file, but let's make sure
it's properly cleaned up after the avocado-vt execution
ends.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>