virttest.asset: Clean test provider directory in case of inconsistency
Fixes issue #126.

Some times, we may press Ctrl+C in the middle of the
bootstrap process (test provider download stage). This
may leave an inconsistent test provider directory.

So let's clean up the provider directory if we capture
a KeyboardInterrupt, or something goes wrong with one
of the git commands executed.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>