virttest: Replace uses of /tmp with data_dir.get_tmp_dir()
Across avocado-vt, we use the /tmp dir extensively. The
problem with that is that many of our files have predictable
names. The predictable names can be used to exploit the system.

One simple solution that can be employed is to use the
securely created avocado temporary directory to store those
files. As the base dir is not predictable, we solve the problem.

So I went through the virttest library code and replaced
all uses of the literal /tmp dir and replaced them with
virttest.data_dir.get_tmp_dir().

About the 'global' VM file locks and address pool locks,
we still do not claim to be fully multi avocado process
run safe yet, so we can get away with moving all those
files to the avocado tmp directory until we figure out
a safe global locking mechanism that can be used for
multiple concurrent avocado executions.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>