unattended_install.py: Bridge IP auto detection for content server
When using the http_ks variants of the unattended install
test, avocado-vt can detect the IP of the main netdst (by
default, virbr0). So instead of leaving an old hardcoded
value in base.cfg, remove it and fix the logic to ensure
the auto detected value is picked over empty values
present in the configuration file.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>