avocado-vt: Move environment file to job temporary directory
Now that the environment file is not persisted between
jobs, it makes sense making it a per job ID instance,
in the avocado temporary directory. This will solve
the problem of stale environment files giving failures
to the user if Ctrl+C was pressed at a moment such that
the env file cleanup did not work.

An alternate solution, removing the env file persistence,
was studied, however, the problem with that line is that
the environment has to persist between each vt test
subprocess, and eliminating the disk representation would
prevent that.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>