Revert "virttest.utils_env: Remove env file upon garbage collection"
Each avocado-vt test will run on a subprocess, and that means
that the environment file will be destroyed at the end of a
test, destroying any living vms in there, regardless of the
value of keep_guest_running in the config file. This was an
oversight, and recently a patch to keep the env file in
the avocado temporary dir has been accepted, which supersedes
this patch.

So let's revert commit 7aeb3c8618e19d28ff6c62103f0b8792961b29e5.

Signed-off-by: Lucas Meneghel Rodrigues <lookkas@gmail.com>