Default guest OS and variant update
Avocado-vt used to give different (older) versions of the JeOS to host
systems that were not a very recent Fedora. This seemed to be
necessary, but those days are gone.

The new unique guest OS and variant will be given to all host systems,
and besides the latest Fedoras which were already using it, it has
been tested under EL6 and Ubuntu 14.04 LTS.

Signed-off-by: Cleber Rosa <crosa@redhat.com>