virttest.qemu_devices.qcontainer: Don't set_clean() when hotplug fails
In simplehotplug, when we can't decide whether the actual hotplug passed
or failed and the qdev representation fails to inject the device,
we set the representation as clean. This might or might not be true,
because the hotplug might have also passed. Therefor this patch removes
the `set_clean()` and simply raise the DeviceHotplugError and leave it
up to the users to decide whether the qdev is in sync or not.

Additionally add support to pass the `ver_out` (output from the qemu
monitor) to the exception so the test writers can read it and decide
whether the device was or was not inserted.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>