docs: DownloadableImages.rst - Update info on the JeOS download location
The new host for the avocado project assets does not allow
for inspecting contents of folders - that is, if you go
to http://assets-avocadoproject.rhcloud.com/static/, you'll
get a 404 error. Update the docs with that limitation (or
feature, depending on how you see it), and pass full urls
to the JeOS images and SHA1SUM files.

Signed-off-by: Lucas Meneghel Rodrigues <lookkas@gmail.com>