Makefile: limit link/unlink targets to config files
In a typical developer workflow, avocado will be run out of the
source tree. For that we used to maintain `link` and `unlink`
targets, which would create symlinks for the libraries, plugin
modules (specific files) and plugins specific configuration files.

Now that the libraries live on their own namespaces, and not anymore
under "avocado" (say avocado.virt), there's no need to link that.

And now plugins are activated by registering setuptools entry points
(please remember to run `$ python setup develop`) that is also not
needed anymore.

Only the configuration files are still handy to have linked.

Signed-off-by: Cleber Rosa <crosa@redhat.com>