vt: Avocado-vt is not enabled when qemu-bin is not found
During plugin initialization we look-out the qemu-bin. There was an
incorrect exception caught to ignore failure which this patch fixes.

Additionally the default value was "Could not find one", which fails
when we try to run avocado with custom qemu (using --vt-qemu-bin)
because obviously this path does not exists. This patch supplies None
and only sets this value in the help text.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>