Guest OS support: add Fedora 22 (x86_64)
Since development related packages are no longer present in the
"Server" ISO and repo, the installation of the c-development
development-tools groups are done in the %post section of the
kickstart file.

It may fail due to lack of connectivity from within the guest, but
it's the best we can do.

The resulting guest has been tested with the
io-github-autotest-qemu.autotest.bonnie test and succeeds.

Signed-off-by: Cleber Rosa <crosa@redhat.com>