virttest/data_dir.py: always sync the download dir
It turns out that the creation of the download dir is done lazily, on
demand. Ideally, that kind of full directory copy would not be done at
a function such as `get_download_dir()`.

While not changing the overall behavior of the current implementation
(IMHO this is needed, but it's a completely different discussion),
let's always sync the contents of those directories.

The use case is that, after an upgrade that adds a new asset (.ini
file) to the download directory, the user won't be able to find the
mentioned asset and will be unable to download it. When the new file
is, say, a new default guest OS file, it resutls in major user
experience breakage.

Also fix the one instance where get_download_dir() was being used
outside of the bootstrap code flow. It was used as a temporary
download location. Yes, the name is rather misleading.

Signed-off-by: Cleber Rosa <crosa@redhat.com>