Job lock: plugin that prevents multiple runs of VT jobs
This introduces a Pre/Post Job type of plugin, that attempts to create
a lock file in a common location before the job is started, and remove
the lock file after the job has finished.  The goal is to prevent
multiple runs of avocado (with VT tests) to corrupt files such as
guest OS images.

It's still possible to have multiple parallel jobs running on the same
machine, as long as their configuration point to different lock
directories, which should follow different data directories, to
prevent the original problem of data file corruption.

The mechanism used for locking is a simplistic one, similar to daemon
PID files.  It should be possible to either rewrite or improve the
lock mechanism completely without bringing changes to the user
expected behavior of, by default, not having multiple Avocado-VT jobs
run at once.

While testing for parallel execution of VT and non-VT jobs, it was
found that the VT loader had a number of issues that make it fail to
run in parallel.  Thse have been fixed and the attempt to execute jobs
in parallel works as expected.

References: https://trello.com/c/hxqiVZCy
References: https://trello.com/c/7xEPKDsJ
Signed-off-by: Cleber Rosa <crosa@redhat.com>