Default guest type: switch to JeOS.23
This adds a new download asset definition for JeOS.23 and switches
the default guest type to that same OS.

Also, it removes the download asset definition for JeOS.20, since it
has actually being removed from the hosting server. The reasons are,
besides having somewhat short storage spaces on the server, we also
don't want to encourage/support the use of such an old version of it.
If users really care about it, they're free to recreate it with the
provided unattended install test and kickstart file.

Signed-off-by: Cleber Rosa <crosa@redhat.com>