avocado_vt: Move the special name handling to loader
Avocado-vt requires a special test-name handling. Currently it does that
inside the `loader` and then it drops it only to evaluate it again. This
patch moves the `VirtTest.__init__` name handling into the `loader` and
replaces the non-used handling there.

This should not have any impact on the test name whatsoever as the
original name was not used at all and the loader now creates the name
according to the originally used `VirtTest.__init__` handling, with the
only difference that the test is set earlier in the process.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>