avocado_vt: Avoid re-creating subtests/guest-os on run
Currently the `subtests.cfg` and `guest-os.cfg` files are always re-
created on `vt-bootstrap` and during each invocation of `avocado-vt`
loader (during `avocado list` and `avocado run`), which is not optimal
and wrong for several reasons.

1. `avocado run` should not modify any configs, it should only run test
2. parallel execution of `avocado run`/`avocado list` causes issues
3. even documentation refers to update on `vt-bootstrap` on some places

This behavior is convenient for test developers, but it does not belongs
to production environment. So this patch improves the documentation and
removes the `subtests.cfg`/`guest-os.cfg` recreation from the loader.

Everysince this commit test developers need to run `avocado vt-
bootstrap` to propagate the changes to subtests/guest-os!

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>