virttest.asset: Avoid providers update on asset import
The test providers are defined in ini files stored in user's datadir.
For convenience these were synced from the avocado base dir on every
`virttest.asset` import, which is unexpected as only bootstrap should
update data dir. This patch moves this update to the bootstrap instead.
Note that the ini files are synced even when the providers are not
downloaded (as before).

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>