virttest.bootstrap: Avoid config_filter caching on import
The `config_filter` was set on import using `_get_config_filter`
function. The problem is, when the directory does not exists yet, then
this filter is incorrect and files like `test-shared.cfg` are included
in the generated `subtests.cfg` (because they are not filtered-out using
the `config_filter`).

This patch removes the cached `config_filter` and renames the
`_get_config_filter` function to `get_config_filter`, which is then used
in functions to get the filter at the time of the function invocation,
which should give the correct results.

This patch fixes the issue where `avocado vt-bootstrap --vt-type spice
--vt-no-downloads --vt-update-config` on empty `~avocado/data` generates
incorrect `subtests.cfg`.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>