Job Lock: swap the logic of lock creation/check
And with that, add support for dealing with multiple lock files, and
also cleaning stale lock files.  After the job has been started, and
it's identified as a VT job, the lock creation logic now consists of:

 1. Create the lock file
 2. Lock files exist?
 2.1. Yes -> Are lock files stale (process do not exist)?
 2.1.1. Yes -> Remove them
 2.1.2. No  -> Abort job
 3. Remove lock file

Some utility methods have been cleaned up, such as _abort() and
_unlock(), which with the changes were being called from on a single
place.

This requires a fix to `avocado.utils.process.pid_exists()` that has
been proposed in the corresponding repository.

Signed-off-by: Cleber Rosa <crosa@redhat.com>