Job Lock: change behavior of the job lock creation utility
This makes the job lock creation utility method do just that: create
the lock.  It doesn't try to protect against possible errors and abort
the job.  This is left only to the try/except block on pre().

Because of that, it raises one specific type of exception when a
failure to create a lock occurs.

Signed-off-by: Cleber Rosa <crosa@redhat.com>