avocado_vt.test: Treat env.save exception as warning
The `env.save` is currently only useful when `keep_guests_running` is
set to allow passing env from test to the following test to speedup the
setup phase. On the other hand this feature causes lots of troubles when
users use unpickable objects in environment.

To balance this this patch treats `env.save` failures as warnings only,
informing about the problematic item. When the last `env.save` fails it
also cleanups the environment to avoid left-over machines.

Note that disabling `env.save` is not really an option as this feature
can be also enabled even per-vm from cartesian_config or by modifying
virttest's params from test (kill_vm = no).

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>