avocado_vt: Make vt_joblock more lenient
The locks are usually created in TMP directory, which often does not
allow removing other users files, which makes the vt_joblock fail even
on non-existing pids. This patch only logs a warning in such case and
proceeds.

Note the process still fails when it's unable to read the file or when
the file does not contain pid.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>