scripts/package_jeos.py: add compatibility flags to qemu-img
Older versions of QEMU are be able to use Qcow2 images created by
newer versions.  That was the case with the first version of the
JeOS.23 image, which has now been updated with with an image that was
created with "0.10" compatibility level.

Let's make that the norm for newly created JeOS images, by means of
this utility script.

Reference: https://trello.com/c/1RlcP6Ji
Signed-off-by: Cleber Rosa <crosa@redhat.com>