virttest.libvirt_vm: Disable display when vga is disabled
The "vga" is used mainly in qemu to define the guest graphic card but
additionally it's used to detect whether machine supports graphic or
not. For example on aarch64 or s390 no gpu is supported and it fails to
start when --vnc is specified. The same logic is also used in kickstarts
to chose text/graphical installation, so let's addopt this policy in
libvirt as well.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>