virttest.libvirt_vm: Support for undefining domains with nvram
Aarch64 requires nvram and currently avocado fails as `virsh undefine`
fails. The https://bugzilla.redhat.com/show_bug.cgi?id=1175120 suggest
it's an expected behavior and we need to use `--nvram` to allow removing
domains with nvrams.

This patch adds generic "options" to the `libvirt_vm.VM.undefine` and
adds the `--nvram` option in `libvirt_vm.VM.undefine` when
`vir_domain_undefine_nvram` is `yes` in the current vm's params. This
way the `--nvram` is automatically added when `VM.remove()` is called,
while allowing to skip it when calling manually `VM.undefine()`.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>