virttest.qemu_vm: Detect virtio-blk driver in qemu_vm
Various machines often require to alter the device models or even types.
Previously this was handled by `guest-hw.cfg` where the models were
hardcoded. This has been working nicely for qemu, but libvirt expects
"virtio" and it does the "virtio" => "virtio-blk-{pci,device}"
internally and it fails when "virtio-blk-device" is defined explicitly.

This patch makes "qemu_vm" acts more like libvirt and it expects only
"virtio" for all kind of machines and based on the conditions there it
overrides it to the desired devices. This way the "virtio" drive_format
is automatically adjusted for qemu when using `image_define_by_params`
(by default) but it leaves the "virtio" type for libvirt.

Due to nature this also fixes the "virtio-scsi-pci" => "virtio-scsi-
device" in the "virtio-scsi" drive_format for mmio-based machines.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>