avocado_vt: Don't quit on get_extra_listing
Currently get_extra_listing quits when it finishes the list_guests,
which is not correct. No plugin should quit avocado directly, it can
only report failure or finish.

This affects the "avocado list --vt-list-guests" which does not lists
only the guests and finish, but it lists both, guests and tests. I
believe this is correct behavior, but if we decide tests should not be
listed when extra_listing is used we can add support for returns to
`get_extra_listings` in avocado.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>