shared.scripts.virtio_console_guest: Wait for sigio signals
The get_sigio_poll_return currently reports immediately, which might not
be desired as signals are asynchronous and there might be delay between
the action and the signal being emitted. This patch uses hardcoded 5s
interval to wait if the expected signals are emitted or it fails with
details.

This solves a real problem on ppc64le, which sometimes delivers the
POLLOUT after the check. With this patch it works well.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>