avocado_vt: Add way to list architectures/machine_types
The cartesian_config knows nothing about machine_types, architectures
and tests, therefor it's impossible to really list the architectures or
machine_types. But it is possible to parse only some of the files
related to information we want to get and filter it accordingly, which
was used in this patch.

The way it works is it parses only "machines.cfg" and "guest-os.cfg",
while it ignores the --vt-machine-type and --vt-arch. The resulting
dictionary should store all available profiles, which happens to be the
combinations of guest-os, machine_type and arch.

Given the nature it's possible to list all variants when supplying
`--vt-guest-os ''` which disables filtering by guest-os.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>