shared.unattended.JeOS-23.ks: Disable selinux
The `selinux-policy` is not installed, while the `selinux` is set as
enabled. For some reason this works on Fedora.23, but it's quite
dangerous running with selinux enabled without policies defined.

My motivation is that I use this KS to create RHEL.7 kickstart and the
latest builds started failing with "Failed to load SELinux policy.".

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>