avocado_vt: Don't override nettype=none
By default nettype value from params is overridden by per-execution
value from config (net_type). The supported options are "user", "bridge"
and "none", where the "none" is not to set the type to none, but to get
the option from cartesian (test config). This patch removes the
cartesian assignment, which returns the correct handling.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>