Added check for SMT state in power system before test run
For KVM to work in POWER8 systems we need to have SMT=off and it is common
that we have it SMT=on by default in all the distributions.

This patch checks whether we are satisfying that condition for POWER8,
if not it will make it off, otherwise errors out.

Signed-off-by: Balamuruhan S <bala24@linux.vnet.ibm.com>
Signed-off-by: Satheesh Rajendran <sathnaga@linux.vnet.ibm.com>