avocado_vt: Change default for --vt-qemu-bin
The `vt_qemu_bin` defines which qemu binary will be used in testing. It
has several special handlings to allow various manipulations, one of
them is that when `vt_qemu_bin` is None and `vt_config` is used, it
won't try to determine `qemu_binary` related params in order to allow
cartesian config to set those.

The problem is, that when `vt_qemu_bin` default is the current qemu
path, it won't allow users to set it to None, therefor this feature is
unavailable and the cartesian config params are overridden by `options`,
which might not be desired.

Even by setting the default to `None` the `qemu_binary` will be
processed and suitable binary found, therefor in reality the default is
still the same, only set later in the code (unless `--vt-qemu-bin` is
used).

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>