cartesian: Parsing suffix before yielding dict
The current Cartesian parsing performance is very low. It'll take
several minutes to list all the cases in a test provider.

After some CPU performance profiling, we can find the bottleneck of
current Cartesian parsing process is a function named postfix_parse.
This function is used to clip the range of a value if its key ends with
'_min', '_max' or '_fixed' and is called after the whole configuration
is combined and generated. Since the combination may generates very huge
number possible results, this function then will be called too many
times.

This patch reduce the calculation by calling this function right before
parser reach a leaf node. This will greatly decrease the number of
calling this function thus increase the performance of list large number
of test cases.

Signed-off-by: Hao Liu <hliu@redhat.com>