virttest.qemu_monitor: Avoid circular dependency
The `Monitor` object is usually part of `VM`, but at the same time it
requires `VM` in order to be created. This brings circular dependency
which breaks pickle makes it impossible to load the state of the running
VM. There are two ways, one is to clean `Monitor` to not contain `VM`,
which would be cleaner, but it'd require changes to all places where
`Monitor` is created and as this would be a very intrusive change
probably breaking many downstream versions, let's simply fake the VM
object and create a dummy class with the important information (the
name) stored there. It's not nice, but it allows keeping the API
unchanged.

This patch is necessary in order to be able to load the env file with a
running VM, eg. for keep_guest_running feature, or when cleaning-up is
disabled and the test fails. For example when `boot.py` test raises
exception and is executed by:

    avocado --show all run boot --vt-extra-params 'kill_vm=no' \
    'kill_vm_on_error=no'

(reported by Andrei Stepanov)

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>