Guest OS Support: add Fedora 25 (x86_64)
This introduces the Fedora 25 (for the x86_64 arch).  This has
been tested with both "cdrom" and "url" methods, and with the
delivery of the "unattended file" (the kickstart file) with
"extra_cdrom_ks" and also the builtin HTTP server.

One particular difference in the kickstart file is that it
includes the installation of Avocado from the official repos.
This may open the way for other types of tests and uses for images
generated with this guest OS type.

Signed-off-by: Cleber Rosa <crosa@redhat.com>