shared.unattended: Rename RHEL.7.devel's ks file
The RHEL.7.devel kickstart is being used both as spice and as normal
kickstart. The big issue is that spice team required some changes and
they sneaked-in a patch which makes kickstarts from "test-providers.d"
taking precedence over the stock ones and later they altered the
"RHEL-7-spice.ks" file in their repo. This is very hard to debug as
without "tp-spice" repo cloned avocado-vt uses the stock RHEL-7-spice.ks
but with "tp-spice" cloned it uses RHEL-7-spice.ks from tp-spice
provider no matter which tp backend is used.

I think this behavior is actually a bug, we should not look for any
config files in backend that are not being used and I want to address
that but in the mean time let's simply rename the RHEL.7.devel's
kicstart to a different name as the current behavior heavily affects
users.

Note the tp-spice is cloned automatically on `avocado vt-bootstrap`
that's what makes this issue pretty hot.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>