avocado_vt: Avoid custom runTest handling
For quite a while we had been using custom runTest handling which copied
most of the Avocado's runTest only with the support to skip the already
running test. This is not really necessary as we can simply execute the
test in setUp phase, which allows the skipping and don't do anything
during test method (as the test already finished)

This version behaves as the old version only by using the Avocado's run
method the whiteboard (and probably other) bugfixes which were applied
to Avocado but not to Avocado-vt are now present.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>