avocado-plugins-vt.spec: allow Avocado LTS version (or later)
Some users have reported that it's impossible to use the current
Avocado-VT packages with an Avocado LTS version.  This is actually a
valid installation scenario, but one that we've so far only supported
with Avocado-VT from GIT sources + Avocado from RPM packages.

One way to fix this is to require *either* the latest Avocado version,
or the LTS version.  The other one is to allow for any version starting
from the (latest) LTS version.  This change uses the later approach,
which seems to give more flexibility to users.  The recommended
installation, though, is either:

 * Avocado-VT (latest) + Avocado (latest)
 * Avocado-VT (latest) + Avocado (LTS)

Signed-off-by: Cleber Rosa <crosa@redhat.com>