avocado-plugins-vt.spec: introduced subpackage qemu
This is an experiment, albeit a functional one, to put test providers
and its dependecies on separate packages.

This came out of users usability comments/requests when bootstrapping
Avocado + Avocado-VT.

By installing avocado-plugins-vt-qemu, the bootstrap process then
knows about the tp-qemu test provider, and its package dependencies.
The idea is to do the same to other test providers.

One point that is particularly open to discussion is putting backend
related libraries (virttest/qemu*.py) in this package.  It can either
be seen as a logical step, or as a violation of the collection of
virttest APIs that are available to any test (not only tp-qemu).

Signed-off-by: Cleber Rosa <crosa@redhat.com>