avocado_vt: Fix the test statuses
Recent changes broke the reported test statuses. This patch keeps the
execution inside setUp phase to be able to skip tests when executing
test using old or older Avocado, but report non-skip results (or all
results for new Avocado) in runTest. This way the reported results
should be always correct.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>